+++ modules/apps/collaboration/document-library/document-library-repository-cmis-impl/src/main/java/com/liferay/document/library/repository/cmis/internal/model/CMISFileEntry.java
@@ -21,6 +21,7 @@ import com.liferay.document.library.kernel.service.DLAppHelperLocalServiceUtil;
 import com.liferay.document.library.kernel.util.DLUtil;
 import com.liferay.document.library.repository.cmis.internal.CMISRepository;
 import com.liferay.exportimport.kernel.lar.StagedModelType;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.lock.Lock;
@@ -46,7 +47,6 @@ import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.MimeTypesUtil;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.io.InputStream;
@@ -55,6 +55,7 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
+import java.util.GregorianCalendar;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -172,7 +173,9 @@ public class CMISFileEntry extends CMISModel implements FileEntry {
 
 	@Override
 	public Date getCreateDate() {
-		return _document.getCreationDate().getTime();
+		GregorianCalendar creationDate = _document.getCreationDate();
+
+		return creationDate.getTime();
 	}
 
 	@Override
@@ -418,7 +421,10 @@ public class CMISFileEntry extends CMISModel implements FileEntry {
 
 	@Override
 	public Date getModifiedDate() {
-		return _document.getLastModificationDate().getTime();
+		GregorianCalendar lastModificationDate =
+			_document.getLastModificationDate();
+
+		return lastModificationDate.getTime();
 	}
 
 	@Override