+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/java/com/liferay/dynamic/data/mapping/form/web/internal/notification/DDMFormEmailNotificationSender.java
@@ -57,6 +57,7 @@ import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.template.soy.utils.SoyHTMLContextValue;
 import com.liferay.portal.util.PrefsPropsUtil;
 
 import java.io.Writer;
@@ -279,7 +280,7 @@ public class DDMFormEmailNotificationSender {
 		}
 
 		fieldMap.put("label", labelString);
-		fieldMap.put("value", sb.toString());
+		fieldMap.put("value", new SoyHTMLContextValue(sb.toString()));
 
 		return fieldMap;
 	}
@@ -416,7 +417,8 @@ public class DDMFormEmailNotificationSender {
 		return LanguageUtil.get(resourceBundle, "someone");
 	}
 
-	protected String getViewFormEntriesURL(DDMFormInstance ddmFormInstance)
+	protected String getViewFormEntriesURL(
+			DDMFormInstance ddmFormInstance, ThemeDisplay themeDisplay)
 		throws PortalException {
 
 		Map<String, String[]> params = new HashMap<>();
@@ -431,14 +433,15 @@ public class DDMFormEmailNotificationSender {
 			portletNamespace.concat("formInstanceId"),
 			new String[] {String.valueOf(ddmFormInstance.getFormInstanceId())});
 
-		return _portal.getControlPanelFullURL(
-			ddmFormInstance.getGroupId(),
-			DDMFormPortletKeys.DYNAMIC_DATA_MAPPING_FORM_ADMIN, params);
+		return _portal.getSiteAdminURL(
+			themeDisplay, DDMFormPortletKeys.DYNAMIC_DATA_MAPPING_FORM_ADMIN,
+			params);
 	}
 
 	protected String getViewFormURL(
 			DDMFormInstance ddmFormInstance,
-			DDMFormInstanceRecord ddmFormInstanceRecord)
+			DDMFormInstanceRecord ddmFormInstanceRecord,
+			ThemeDisplay themeDisplay)
 		throws PortalException {
 
 		Map<String, String[]> params = new HashMap<>();
@@ -458,9 +461,9 @@ public class DDMFormEmailNotificationSender {
 			portletNamespace.concat("formInstanceId"),
 			new String[] {String.valueOf(ddmFormInstance.getFormInstanceId())});
 
-		return _portal.getControlPanelFullURL(
-			ddmFormInstance.getGroupId(),
-			DDMFormPortletKeys.DYNAMIC_DATA_MAPPING_FORM_ADMIN, params);
+		return _portal.getSiteAdminURL(
+			themeDisplay, DDMFormPortletKeys.DYNAMIC_DATA_MAPPING_FORM_ADMIN,
+			params);
 	}
 
 	protected void populateParameters(
@@ -477,11 +480,16 @@ public class DDMFormEmailNotificationSender {
 		template.put(
 			"siteName", getSiteName(ddmFormInstance.getGroupId(), locale));
 		template.put("userName", getUserName(ddmFormInstanceRecord, locale));
+
+		ThemeDisplay themeDisplay = getThemeDisplay(portletRequest);
+
 		template.put(
-			"viewFormEntriesURL", getViewFormEntriesURL(ddmFormInstance));
+			"viewFormEntriesURL",
+			getViewFormEntriesURL(ddmFormInstance, themeDisplay));
 		template.put(
 			"viewFormURL",
-			getViewFormURL(ddmFormInstance, ddmFormInstanceRecord));
+			getViewFormURL(
+				ddmFormInstance, ddmFormInstanceRecord, themeDisplay));
 	}
 
 	protected String render(Template template) throws TemplateException {