+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/internal/SoyTemplate.java
@@ -24,6 +24,8 @@ import com.google.template.soy.tofu.SoyTofu;
 import com.google.template.soy.tofu.SoyTofu.Renderer;
 import com.google.template.soy.tofu.SoyTofuOptions;
 
+import com.liferay.portal.kernel.bean.BeanPropertiesUtil;
+import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -48,6 +50,8 @@ import com.liferay.portal.template.soy.utils.SoyTemplateResourcesProvider;
 import java.io.Reader;
 import java.io.Writer;
 
+import java.lang.reflect.Array;
+
 import java.security.AccessController;
 import java.security.PrivilegedActionException;
 import java.security.PrivilegedExceptionAction;
@@ -55,6 +59,7 @@ import java.security.PrivilegedExceptionAction;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -62,8 +67,13 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.ResourceBundle;
 import java.util.Set;
+import java.util.TreeMap;
 
 import javax.servlet.http.HttpServletRequest;
+
+import org.apache.commons.lang3.ClassUtils;
+
+import org.json.JSONArray;
 
 import org.osgi.framework.Bundle;
 import org.osgi.framework.wiring.BundleWiring;
@@ -209,26 +219,136 @@ public class SoyTemplate extends AbstractMultiResourceTemplate {
 	}
 
 	protected Object getSoyMapValue(Object value) {
-		Object soyMapValue = null;
-
 		if (value == null) {
-			soyMapValue = null;
+			return null;
 		}
-		else if (value instanceof SoyHTMLContextValue) {
+
+		Class<?> clazz = value.getClass();
+
+		if (ClassUtils.isPrimitiveOrWrapper(clazz) || value instanceof String) {
+			return value;
+		}
+
+		if (clazz.isArray()) {
+			List<Object> newList = new ArrayList<>();
+
+			for (int i = 0; i < Array.getLength(value); i++) {
+				Object object = Array.get(value, i);
+
+				newList.add(getSoyMapValue(object));
+			}
+
+			return newList;
+		}
+
+		if (value instanceof Iterable) {
+			@SuppressWarnings("unchecked")
+			Iterable<Object> iterable = (Iterable<Object>)value;
+
+			List<Object> newList = new ArrayList<>();
+
+			for (Object object : iterable) {
+				newList.add(getSoyMapValue(object));
+			}
+
+			return newList;
+		}
+
+		if (value instanceof JSONArray) {
+			JSONArray jsonArray = (JSONArray)value;
+
+			List<Object> newList = new ArrayList<>();
+
+			for (int i = 0; i < jsonArray.length(); i++) {
+				Object object = jsonArray.opt(i);
+
+				newList.add(getSoyMapValue(object));
+			}
+
+			return newList;
+		}
+
+		if (value instanceof Map) {
+			Map<Object, Object> map = (Map<Object, Object>)value;
+
+			Map<Object, Object> newMap = new TreeMap<>();
+
+			for (Map.Entry<Object, Object> entry : map.entrySet()) {
+				Object newKey = getSoyMapValue(entry.getKey());
+
+				if (newKey == null) {
+					continue;
+				}
+
+				Object newValue = getSoyMapValue(entry.getValue());
+
+				newMap.put(newKey, newValue);
+			}
+
+			return newMap;
+		}
+
+		if (value instanceof JSONObject) {
+			JSONObject jsonObject = (JSONObject)value;
+
+			Map<String, Object> newMap = new TreeMap<>();
+
+			Iterator<String> iterator = jsonObject.keys();
+
+			while (iterator.hasNext()) {
+				String key = iterator.next();
+
+				Object object = jsonObject.get(key);
+
+				Object newValue = getSoyMapValue(object);
+
+				newMap.put(key, newValue);
+			}
+
+			return newMap;
+		}
+
+		if (value instanceof org.json.JSONObject) {
+			org.json.JSONObject jsonObject = (org.json.JSONObject)value;
+
+			Map<Object, Object> newMap = new TreeMap<>();
+
+			Iterator<String> iterator = jsonObject.keys();
+
+			while (iterator.hasNext()) {
+				String key = iterator.next();
+
+				Object object = jsonObject.opt(key);
+
+				Object newValue = getSoyMapValue(object);
+
+				newMap.put(key, newValue);
+			}
+
+			return newMap;
+		}
+
+		if (value instanceof SoyHTMLContextValue) {
 			SoyHTMLContextValue htmlValue = (SoyHTMLContextValue)value;
 
-			soyMapValue = htmlValue.getValue();
+			return htmlValue.getValue();
 		}
-		else if (value instanceof SoyRawData) {
+
+		if (value instanceof SoyRawData) {
 			SoyRawData soyRawData = (SoyRawData)value;
 
-			soyMapValue = soyRawData.getValue();
-		}
-		else {
-			soyMapValue = _templateContextHelper.deserializeValue(value);
+			return soyRawData.getValue();
 		}
 
-		return soyMapValue;
+		Map<String, Object> newMap = new TreeMap<>();
+
+		BeanPropertiesUtil.copyProperties(value, newMap);
+
+		if (newMap.isEmpty()) {
+			return null;
+		}
+
+		return getSoyMapValue(newMap);
 	}
 
 	protected Optional<SoyMsgBundle> getSoyMsgBundle(