+++ modules/apps/web-experience/asset/asset-browser-web/src/main/java/com/liferay/asset/browser/web/internal/display/context/AssetBrowserDisplayContext.java
@@ -23,7 +23,9 @@ import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil;
 import com.liferay.asset.util.AssetHelper;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.NavigationItem;
 import com.liferay.frontend.taglib.servlet.taglib.ManagementBarFilterItem;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.model.Group;
@@ -37,12 +39,10 @@ import com.liferay.portal.kernel.search.Hits;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.service.GroupLocalServiceUtil;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.HttpUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
@@ -293,38 +293,40 @@ public class AssetBrowserDisplayContext {
 		List<ManagementBarFilterItem> managementBarFilterItems =
 			new ArrayList<>();
 
-		long[] selectedGroupIds = ArrayUtil.append(
-			new long[] {0}, getSelectedGroupIds());
+		String label = LanguageUtil.get(_request, "all");
 
-		for (long curGroupId : selectedGroupIds) {
-			Group curGroup = GroupLocalServiceUtil.fetchGroup(curGroupId);
+		PortletURL groupURL = getPortletURL();
 
-			if ((curGroup == null) && (curGroupId > 0)) {
-				continue;
-			}
+		groupURL.setParameter("groupId", "0");
 
+		ManagementBarFilterItem managementBarFilterItem =
+			new ManagementBarFilterItem(false, label, groupURL.toString());
+
+		managementBarFilterItems.add(managementBarFilterItem);
+
+		long selectedGroupId = ParamUtil.getLong(_request, "selectedGroupId");
+
+		long[] selectedGroupIds = PortalUtil.getSharedContentSiteGroupIds(
+			themeDisplay.getCompanyId(), selectedGroupId,
+			themeDisplay.getUserId());
+
+		List<Group> groups = GroupLocalServiceUtil.getGroups(selectedGroupIds);
+
+		for (Group curGroup : groups) {
 			boolean active = false;
 
-			if (getGroupId() == curGroupId) {
+			if (getGroupId() == curGroup.getGroupId()) {
 				active = true;
 			}
 
-			String label = StringPool.BLANK;
+			label = HtmlUtil.escape(
+				curGroup.getDescriptiveName(themeDisplay.getLocale()));
 
-			if (curGroup != null) {
-				label = HtmlUtil.escape(
-					curGroup.getDescriptiveName(themeDisplay.getLocale()));
-			}
-			else {
-				label = LanguageUtil.get(_request, "all");
-			}
+			groupURL.setParameter(
+				"groupId", String.valueOf(curGroup.getGroupId()));
 
-			PortletURL groupURL = getPortletURL();
-
-			groupURL.setParameter("groupId", String.valueOf(curGroupId));
-
-			ManagementBarFilterItem managementBarFilterItem =
-				new ManagementBarFilterItem(active, label, groupURL.toString());
+			managementBarFilterItem = new ManagementBarFilterItem(
+				active, label, groupURL.toString());
 
 			managementBarFilterItems.add(managementBarFilterItem);
 		}
@@ -343,6 +345,24 @@ public class AssetBrowserDisplayContext {
 		Group group = GroupLocalServiceUtil.fetchGroup(getGroupId());
 
 		return group.getDescriptiveName(themeDisplay.getLocale());
+	}
+
+	public List<NavigationItem> getNavigationItems() throws PortalException {
+		List<NavigationItem> navigationItems = new ArrayList<>();
+
+		NavigationItem entriesNavigationItem = new NavigationItem();
+
+		entriesNavigationItem.setActive(true);
+
+		PortletURL mainURL = getPortletURL();
+
+		entriesNavigationItem.setHref(mainURL.toString());
+
+		entriesNavigationItem.setLabel(LanguageUtil.get(_request, "entries"));
+
+		navigationItems.add(entriesNavigationItem);
+
+		return navigationItems;
 	}
 
 	public String getOrderByCol() {
@@ -378,13 +398,6 @@ public class AssetBrowserDisplayContext {
 		PortletURL portletURL = _renderResponse.createRenderURL();
 
 		portletURL.setParameter("groupId", String.valueOf(getGroupId()));
-
-		long[] selectedGroupIds = getSelectedGroupIds();
-
-		if (selectedGroupIds.length > 0) {
-			portletURL.setParameter(
-				"selectedGroupIds", StringUtil.merge(selectedGroupIds));
-		}
 
 		long selectedGroupId = ParamUtil.getLong(_request, "selectedGroupId");