+++ modules/apps/web-experience/fragment/fragment-service/src/main/java/com/liferay/fragment/model/impl/FragmentEntryImpl.java
@@ -17,6 +17,7 @@ package com.liferay.fragment.model.impl;
 import com.liferay.fragment.model.FragmentEntry;
 import com.liferay.html.preview.model.HtmlPreviewEntry;
 import com.liferay.html.preview.service.HtmlPreviewEntryLocalServiceUtil;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.sanitizer.Sanitizer;
 import com.liferay.portal.kernel.sanitizer.SanitizerUtil;
@@ -25,7 +26,8 @@ import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.StringBundler;
-import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.xml.simple.Element;
+import com.liferay.portal.kernel.zip.ZipWriter;
 
 import java.util.Optional;
 
@@ -78,6 +80,27 @@ public class FragmentEntryImpl extends FragmentEntryBaseImpl {
 		}
 
 		return htmlPreviewEntry.getImagePreviewURL(themeDisplay);
+	}
+
+	@Override
+	public void populateZipWriter(ZipWriter zipWriter, String path)
+		throws Exception {
+
+		path = path + StringPool.SLASH + getFragmentEntryId();
+
+		Element fragmentEntryElement = new Element("fragment-entry", false);
+
+		fragmentEntryElement.addElement("name", getName());
+		fragmentEntryElement.addElement("css-path", path + "/css.css");
+		fragmentEntryElement.addElement("js-path", path + "/js.js");
+		fragmentEntryElement.addElement("html-path", path + "/html.html");
+
+		zipWriter.addEntry(
+			path + "/definition.xml", fragmentEntryElement.toXMLString());
+
+		zipWriter.addEntry(path + "/css.css", getCss());
+		zipWriter.addEntry(path + "/js.js", getJs());
+		zipWriter.addEntry(path + "/html.html", getHtml());
 	}
 
 }