+++ modules/apps/web-experience/site-navigation/site-navigation-menu-item-node/src/main/java/com/liferay/site/navigation/menu/item/node/internal/type/NodeSiteNavigationMenuItemType.java
@@ -15,6 +15,8 @@
 package com.liferay.site.navigation.menu.item.node.internal.type;
 
 import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.util.ResourceBundleLoader;
 import com.liferay.portal.kernel.util.UnicodeProperties;
@@ -74,6 +76,15 @@ public class NodeSiteNavigationMenuItemType
 	@Override
 	public String getType() {
 		return SiteNavigationMenuItemTypeNodeConstants.NODE;
+	}
+
+	@Override
+	public String getURL(
+			HttpServletRequest request,
+			SiteNavigationMenuItem siteNavigationMenuItem)
+		throws PortalException {
+
+		return StringPool.BLANK;
 	}
 
 	@Override